+++ Avalonia.Lottie/LinearInterpolator.cs
@@ -2,9 +2,9 @@ namespace Avalonia.Lottie
 {
     public class LinearInterpolator : IInterpolator
     {
-        public float GetInterpolation(float f)
+        public double  GetInterpolation(double f)
         {
-            if (f < 0 || float.IsNaN(f))
+            if (f < 0 || double .IsNaN(f))
                 f = 0;
             if (f > 1)
                 f = 1;
